requirements.txt
pandas==2.3.2
PyMySQL==1.1.2
python-dotenv==1.1.1
SQLAlchemy==2.0.43
numpy==2.3.3
python-dateutil==2.9.0.post0
typing_extensions==4.15.0